requirement.txt
python=3.11
google.generativeai
